adjust ui create story and story view and fix api create story

+++ lib/util/global.dart
@@ -63,7 +63,7 @@ class Global {
     return false;
   }
 
-  static String convertMedia(String path,double? width,int? height) {
+  static String convertMedia(String path,double? width,double? height) {
     debugPrint("Loaded path: " + path);
     return "http://14.225.204.248:8080/$path?width=$width&height=$height&fit=cover";
   }
